Work on rcopy().
Since rcopy() is implemented in FCR as a series of goto & statements, we can
try that out here as well.  However, we only need to call the internal
argument validation subroutines once.  So we refactor them such they can be
individually called as needed within the three public functions.

However, this is not debugged.  Have TODO-ed one test in t/003-rcopy.t.  Tests
are leaving a plain-text file named 'foo' in current working directory.